Don't use `--create`
In the current post-award(data-store) world, our dbname is
`data_api_db`. In pre-award(post-award) world, our dbname is
`post_award`.

Using the `--create` flag means that we are restoring into target db
under the `data_api_db` name, which is wrong.

If we drop `--create` and just use `--clean`, then everything gets
restored into the database we're connected to - which will be from the
URI, and be the correct (`post_award`) location.