Update fund-store URL for combined pre-award-stores
We are migrating from a bunch of "microservice"-y pre-award API services
to a single combined "pre-award-stores" API.

After we've migrated, we need to point all apps/lambdas/services/etc to
the new service, and the fund-store API endpoints have now been
namespaced under a /fund prefix.

All of the other apps will read the fund-store host from an AWS
parameter store thing, but lambdas aren't able to read this in the same
way - they would be expected to retrieve the value at runtime.

I don't want to change the code to support this, and since this is the
only outlier I'm going to take the simplest approach - just update the
fund store host in a PR. We can manage this manually as we do the
migration/switchover on the day.