Run DB migrations using paketo image
Remove the unused inputs for this workflow, and pass the
`image_location` into the migration scripts (which will be updated in
separate per-repo PRs) so that they can reuse this image rather than
building a new image from the Dockerfile in the repo. This will result
in a more consistent image used for DB changes, and should result in
faster deployments because we don't build the image twice (once in
paketo, once with a Dockerfile).