FSPT-157: remove references to funding service design base repo
The funding service design base repository won't be as relevant now that
are removing the number of repositories and steering away from a
microservice architecture.

It doesn't look like it was ever used by the repositories so remove
references to it before archiving.

Of note while removing this it looks like some of these instructions
will need to change as we move to a single repository - some may already
be out of date so an audit might be useful as a step at the end of this
work.