Opt-in to python tests
These tests can't run on the UAT environment, but with default=true they
will try to.

Let's opt-in to running the python tests instead.

This means we need to update the other repos (pre-award-frontend and
pre-award-stores) but it seems the clearest way to do it.

An alternative would be to keep default=true and then have a conditional
check on the python job to ignore the UAT environment, but that feels
likely to confuse someone in the future (even though we aim to remove
uat soon).