Don't notify slack on rejected pipeline runs
We have deployment protection rules on our UAT/prod environments that
require a developer to go and manually approve deployments to those
environments.

If a deployment is rejected, github - for some reason - records that as
a failed run rather than a cancelled run.

There seems to be no trivial way of retrieving a clean value for a
manual rejection, so there are some workarounds suggested instead:

https://github.com/orgs/community/discussions/66483

While that solution does give access to a clean value that has a strong
correlation with approval/rejection, it breaks down on re-runs.

I think the implementation I've gone for here works better with re-runs,
however it does assert on a plain string rather than a boolean value, so
theoretically it could break down if GitHub changes that string.

If the annotation message _does_ change, this check will fail open and
slack will get notified, which feels safe as fallback behaviour.

Alternatively, if we end up having other protection rules in place that
block deployments beyond manual approval. But we are aiming to head
towards zero-approvals CI
(https://github.com/communitiesuk/funding-service-requests-for-comments/discussions/21).
If/when that happens we can remove this check and accept any failures as
legit ones that should trigger the slack notification.