Concurrency group for each app name
It should be possible to deploy two things at the same time - all that
matters is that we aren't deploying a single thing multiple times at
once.

If we really wanted to only have one service/job/etc deployable a time,
then theoretically we should also include running the test suite as part
of that group. But we don't. So we clearly don't care about that level
of isolation. So I'm removing it.