Delete application deadline reminder lambda
This has now been reproduced in the pre-award repo as a scheduled job:
https://github.com/communitiesuk/funding-service-pre-award/pull/243.

Note: we have not yet reproduced the 'incomplete applications' email,
which this lambda also did.

There is an outstanding product question around whether we bother to
reproduce this or not, as it hasn't been used since it was refactored to
use the events table many months ago.